backend/requirements.txt
PyJWT[crypto]>=2.0.0
cryptography
requests
Django
djangorestframework
dj-rest-auth
django-allauth
django-cors-headers
django-extensions
django-environ
